rename: common images 파일 케밥케이스로 통일
