+++ app/common/assets/images/curation-no-seat.png
