+++ app/common/components/layout/DetectDevice/DeskTopLayout.tsx
@@ -1,6 +1,6 @@
 import dynamic from "next/dynamic";
 const PhoneRingingImage = dynamic(
-  () => import("@/common/assets/images/phoneRinging.svg")
+  () => import("@/common/assets/images/phone-ringing.svg")
 );
 const QRCode = dynamic(() => import("./QRCode"));
